src/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
torch==2.6.0
torchvision==0.21.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.30
Pillow==11.0.0
scikit-image==0.25.2
numpy==2.1.3
scipy==1.14.1
sentence-transformers==3.3.1
transformers==4.47.1
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
gradio>=6.0.0
PyWavelets>=1.6.0
